% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_sizer.R
\name{blog_img_wd}
\alias{blog_img_wd}
\title{Crop and scale header to wide}
\usage{
blog_img_wd(
  image,
  width_px = 800,
  height_px = 200,
  wcrop_px = 5000,
  hcrop_px = 1000,
  woffset_px = 200,
  hoffset_px = 1400
)
}
\arguments{
\item{image}{a jpeg image}

\item{width_px}{output width pixels}

\item{height_px}{output height pixels}

\item{wcrop_px}{selected width pixels to crop, defaults to center?}

\item{hcrop_px}{selected height pixels to crop defaults to center?}

\item{woffset_px}{plus or minus horizontal pixels from center most}

\item{hoffset_px}{plus or minus vertical pixels from center most}
}
\value{

}
\description{
Crop and scale header to wide
}
