% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_img_data.R
\docType{data}
\name{header_img_data}
\alias{header_img_data}
\title{Header Image Data}
\format{
A \code{ggplot} object
}
\usage{
header_img_data
}
\description{
A \code{ggplot} object that is for simple testing and
running of examples for magick processing
}
\keyword{datasets}
